Resize rules are now working properly
