+++ dev/classes/exception.js
